drop `encoding` dependency for UTF16
This causes us to parse UTF-16 rather than UCS-2, i.e. we will incorrectly
accept a few invalid codepoints. Given that the primary use of this library
is ASCII-only, this seems like a good tradeoff for eliminating several very
heavy (parsing) dependencies.

Sets minimum compiler version to 1.15 to allow encoding. I may feature-gate
this in future if it turns out to be the only blocker for 1.14.